+++ pkg/annotation/annotation.go
@@ -1,7 +1,6 @@
 package annotation
 
 const (
-	AzureMachinePoolSubnet   = "azure-machine-pool.giantswarm.io/subnet-name"
 	IsMasterUpgrading        = "azure-machine-pool.giantswarm.io/is-master-upgrading"
 	StateMachineCurrentState = "azure-machine-pool.giantswarm.io/state-machine-current-state"
 )