+++ service/controller/cloudconfig/master_template.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 
 	"github.com/giantswarm/certs"
-	k8scloudconfig "github.com/giantswarm/k8scloudconfig/v6/pkg/template"
+	k8scloudconfig "github.com/giantswarm/k8scloudconfig/v7/pkg/template"
 	"github.com/giantswarm/microerror"
 
 	"github.com/giantswarm/azure-operator/v4/service/controller/encrypter"
@@ -50,8 +50,8 @@ func (c CloudConfig) NewMasterTemplate(ctx context.Context, data IgnitionTemplat
 			azure:                        c.azure,
 			azureClientCredentialsConfig: c.azureClientCredentials,
 			calicoCIDR:                   data.CustomObject.Spec.Azure.VirtualNetwork.CalicoSubnetCIDR,
-			cluster:                      data.CustomObject.Spec.Cluster,
 			clusterCerts:                 data.ClusterCerts,
+			customObject:                 data.CustomObject,
 			encrypter:                    encrypter,
 			subscriptionID:               c.subscriptionID,
 			vnetCIDR:                     data.CustomObject.Spec.Azure.VirtualNetwork.CIDR,
@@ -114,6 +114,8 @@ func (c CloudConfig) NewMasterTemplate(ctx context.Context, data IgnitionTemplat
 			LogsToken:  c.ignition.LogsToken,
 		}
 		params.Images = data.Images
+		params.RegistryMirrors = c.registryMirrors
+		params.Versions = data.Versions
 		params.SSOPublicKey = c.ssoPublicKey
 	}
 	ignitionPath := k8scloudconfig.GetIgnitionPath(c.ignition.Path)