+++ service/controller/cloudconfig/worker_template.go
@@ -5,7 +5,7 @@ import (
 	"encoding/base64"
 
 	"github.com/giantswarm/certs"
-	k8scloudconfig "github.com/giantswarm/k8scloudconfig/v6/pkg/template"
+	k8scloudconfig "github.com/giantswarm/k8scloudconfig/v7/pkg/template"
 	"github.com/giantswarm/microerror"
 
 	"github.com/giantswarm/azure-operator/v4/service/controller/encrypter"
@@ -22,18 +22,11 @@ func (c CloudConfig) NewWorkerTemplate(ctx context.Context, data IgnitionTemplat
 		be := baseExtension{
 			azure:                        c.azure,
 			azureClientCredentialsConfig: c.azureClientCredentials,
-			calicoCIDR:                   "",
-			cluster:                      data.CustomObject.Spec.Cluster,
 			clusterCerts:                 data.ClusterCerts,
-			primaryScaleSetName:          data.PrimaryScaleSetName,
-			resourceGroup:                data.ResourceGroup,
-			routeTableName:               data.RouteTableName,
-			securityGroupName:            data.SecurityGroupName,
-			subnetName:                   data.SubnetName,
+			customObject:                 data.CustomObject,
 			encrypter:                    encrypter,
 			subscriptionID:               c.subscriptionID,
-			vnetCIDR:                     data.VnetCIDR,
-			vnetName:                     data.SubnetName,
+			vnetCIDR:                     data.CustomObject.Spec.Azure.VirtualNetwork.CIDR,
 		}
 
 		params = k8scloudconfig.DefaultParams()
@@ -59,6 +52,8 @@ func (c CloudConfig) NewWorkerTemplate(ctx context.Context, data IgnitionTemplat
 			LogsToken:  c.ignition.LogsToken,
 		}
 		params.Images = data.Images
+		params.RegistryMirrors = c.registryMirrors
+		params.Versions = data.Versions
 		params.SSOPublicKey = c.ssoPublicKey
 
 		ignitionPath := k8scloudconfig.GetIgnitionPath(c.ignition.Path)