+++ service/controller/key/key.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/Azure/go-autorest/autorest/to"
 	providerv1alpha1 "github.com/giantswarm/apiextensions/pkg/apis/provider/v1alpha1"
-	k8scloudconfig "github.com/giantswarm/k8scloudconfig/v6/pkg/template"
+	k8scloudconfig "github.com/giantswarm/k8scloudconfig/v7/pkg/template"
 	"github.com/giantswarm/microerror"
 	capzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
 	expcapzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3"
@@ -60,7 +60,7 @@ const (
 	kubernetesAPIHealthzVersion = "0.1.0"
 
 	// k8s-setup-network-environment.
-	kubernetesNetworkSetupDocker = "0.1.0"
+	kubernetesNetworkSetupDocker = "0.2.0"
 )
 
 var (
@@ -113,8 +113,8 @@ func BlobContainerName() string {
 	return blobContainerName
 }
 
-func BlobName(customObject providerv1alpha1.AzureConfig, role string) string {
-	return fmt.Sprintf("%s-%s-%s", OperatorVersion(&customObject), cloudConfigVersion, role)
+func BlobName(customObject LabelsGetter, role string) string {
+	return fmt.Sprintf("%s-%s-%s", OperatorVersion(customObject), cloudConfigVersion, role)
 }
 
 func WorkerBlobName(operatorVersion string) string {
@@ -125,8 +125,8 @@ func CalicoCIDR(customObject providerv1alpha1.AzureConfig) string {
 	return customObject.Spec.Azure.VirtualNetwork.CalicoSubnetCIDR
 }
 
-func CertificateEncryptionSecretName(customObject providerv1alpha1.AzureConfig) string {
-	return fmt.Sprintf("%s-certificate-encryption", customObject.Spec.Cluster.ID)
+func CertificateEncryptionSecretName(customObject LabelsGetter) string {
+	return fmt.Sprintf("%s-certificate-encryption", ClusterID(customObject))
 }
 
 func CloudConfigSmallTemplates() []string {
@@ -367,13 +367,13 @@ func AvailabilityZones(customObject providerv1alpha1.AzureConfig, location strin
 	return customObject.Spec.Azure.AvailabilityZones
 }
 
-func StorageAccountName(customObject providerv1alpha1.AzureConfig) string {
+func StorageAccountName(customObject LabelsGetter) string {
 	// In integration tests we use hyphens which are not allowed. We also
 	// need to keep the name globaly unique and within 24 character limit.
 	//
 	//	See https://docs.microsoft.com/en-us/azure/architecture/best-practices/naming-conventions#storage
 	//
-	storageAccountName := fmt.Sprintf("%s%s", storageAccountSuffix, ClusterID(&customObject))
+	storageAccountName := fmt.Sprintf("%s%s", storageAccountSuffix, ClusterID(customObject))
 	return strings.Replace(storageAccountName, "-", "", -1)
 }