+++ service/controller/resource/cloudconfig/types.go
@@ -1,0 +1,13 @@
+package cloudconfig
+
+const (
+	prefixMaster = "master"
+	prefixWorker = "worker"
+)
+
+type ContainerObjectState struct {
+	ContainerName      string
+	Body               string
+	Key                string
+	StorageAccountName string
+}