+++ service/service.go
@@ -40,11 +40,12 @@ type Config struct {
 	Flag  *flag.Flag
 	Viper *viper.Viper
 
-	Description string
-	GitCommit   string
-	ProjectName string
-	Source      string
-	Version     string
+	Description    string
+	GitCommit      string
+	ProjectName    string
+	RegistryDomain string
+	Source         string
+	Version        string
 }
 
 type Service struct {
@@ -217,7 +218,7 @@ func New(config Config) (*Service, error) {
 			OIDC:             OIDC,
 			InstallationName: config.Viper.GetString(config.Flag.Service.Installation.Name),
 			ProjectName:      config.ProjectName,
-			RegistryDomain:   config.Viper.GetString(config.Flag.Service.RegistryDomain),
+			RegistryDomain:   config.Viper.GetString(config.Flag.Service.Registry.Domain),
 			SSOPublicKey:     config.Viper.GetString(config.Flag.Service.Tenant.SSH.SSOPublicKey),
 			VMSSCheckWorkers: config.Viper.GetInt(config.Flag.Service.Azure.VMSSCheckWorkers),
 
@@ -285,7 +286,8 @@ func New(config Config) (*Service, error) {
 			Logger:                    config.Logger,
 			OIDC:                      OIDC,
 			ProjectName:               config.ProjectName,
-			RegistryDomain:            config.Viper.GetString(config.Flag.Service.RegistryDomain),
+			RegistryDomain:            config.Viper.GetString(config.Flag.Service.Registry.Domain),
+			RegistryMirrors:           config.Viper.GetStringSlice(config.Flag.Service.Registry.Mirrors),
 			SentryDSN:                 sentryDSN,
 			SSOPublicKey:              config.Viper.GetString(config.Flag.Service.Tenant.SSH.SSOPublicKey),
 			VMSSCheckWorkers:          config.Viper.GetInt(config.Flag.Service.Azure.VMSSCheckWorkers),
@@ -309,6 +311,7 @@ func New(config Config) (*Service, error) {
 			K8sClient:                 k8sClient,
 			Locker:                    kubeLockLocker,
 			Logger:                    config.Logger,
+			RegistryDomain:            config.Viper.GetString(config.Flag.Service.Registry.Domain),
 			SentryDSN:                 sentryDSN,
 			VMSSCheckWorkers:          config.Viper.GetInt(config.Flag.Service.Azure.VMSSCheckWorkers),
 			VMSSMSIEnabled:            config.Viper.GetBool(config.Flag.Service.Azure.MSI.Enabled),