+++ CHANGELOG.md
@@ -12,16 +12,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Mapping from Cluster API & CAPZ CRs to AzureConfig. This change provides migration path towards Azure Cluster API implementation.
 - State machine flowchart generation.
 - Support to forward errors to Sentry.
+- New `cloudconfig` handler for the `AzureCluster` controller that creates the required cloudconfig files in the Storage Account.
+- Add --service.registry.mirrors flag for setting registry mirror domains.
 
 ### Changed
 
+- Allow tenant cluster to be created without built-in workers.
 - Changed how the Azure authentication works when connecting to a different Subscription than the Control Plane's one.
 - Restricted storage account access to the local VNET only.
 - Removed the flatcar migration state machine transitions.
 - Calculate CIDR for a new Tenant Cluster using a local resource rather than getting it from `kubernetesd`.
 - Migrate the `vmsscheck` guards to use the Azure client factory.
 - Use `0.1.0` tag for `k8s-api-heahtz` image.
-- Use `0.1.0` tag for `k8s-setup-network-env` image.
+- Use `0.2.0` tag for `k8s-setup-network-env` image.
+- Replace --service.registrydomain with --service.registry.domain.
 
 ### Removed